Hola

+++ challenge/inverse.h
@@ -4,7 +4,7 @@ void getCofactor(double** A, double** temp, int p, int q, int n, int m);
  
 /* Recursive function for finding determinant of matrix.
    n is current dimension of A[][]. */
-int determinant(double** A, int n,int m);
+double determinant(double** A, int n,int m);
  
 // Function to get adjoint of A[N][N] in adj[N][N].
 void adjoint(double** A,double** adj,int n, int m);
